Fix some warnings

+++ retail/bootloader/source/arm7/patch_common.c
@@ -8517,15 +8517,15 @@ void patchDSiModeToDSMode(cardengineArm9* ce9, const tNDSHeader* ndsHeader) {
 	else if (strcmp(romTid, "KAUE") == 0 && debugOrMep) {
 		extern u32* nintCdwnCalHeapAlloc;
 
-		setBL(0x02012480, dsiSaveGetLength);
-		setBL(0x020124C0, dsiSaveRead);
-		setBL(0x0201253C, dsiSaveWrite);
-		setBL(0x02012B20, dsiSaveOpen);
-		setBL(0x02012BA0, dsiSaveClose);
-		setBL(0x02012F40, dsiSaveCreate);
-		setBL(0x02012F50, dsiSaveOpen);
-		setBL(0x02012F64, dsiSaveSetLength);
-		setBL(0x02012FAC, dsiSaveClose);
+		setBL(0x02012480, (u32)dsiSaveGetLength);
+		setBL(0x020124C0, (u32)dsiSaveRead);
+		setBL(0x0201253C, (u32)dsiSaveWrite);
+		setBL(0x02012B20, (u32)dsiSaveOpen);
+		setBL(0x02012BA0, (u32)dsiSaveClose);
+		setBL(0x02012F40, (u32)dsiSaveCreate);
+		setBL(0x02012F50, (u32)dsiSaveOpen);
+		setBL(0x02012F64, (u32)dsiSaveSetLength);
+		setBL(0x02012FAC, (u32)dsiSaveClose);
 		*(u32*)0x02013124 = 0xE1A00000; // nop
 		*(u32*)0x0201488C = 0xE1A00000; // nop
 		*(u32*)0x02014894 = 0xE1A00000; // nop
@@ -8551,15 +8551,15 @@ void patchDSiModeToDSMode(cardengineArm9* ce9, const tNDSHeader* ndsHeader) {
 	else if (strcmp(romTid, "KAUV") == 0 && debugOrMep) {
 		extern u32* nintCdwnCalHeapAlloc;
 
-		setBL(0x020124DC, dsiSaveGetLength);
-		setBL(0x0201251C, dsiSaveRead);
-		setBL(0x02012598, dsiSaveWrite);
-		setBL(0x02012B7C, dsiSaveOpen);
-		setBL(0x02012BFC, dsiSaveClose);
-		setBL(0x02012F9C, dsiSaveCreate);
-		setBL(0x02012FAC, dsiSaveOpen);
-		setBL(0x02012FC0, dsiSaveSetLength);
-		setBL(0x02013008, dsiSaveClose);
+		setBL(0x020124DC, (u32)dsiSaveGetLength);
+		setBL(0x0201251C, (u32)dsiSaveRead);
+		setBL(0x02012598, (u32)dsiSaveWrite);
+		setBL(0x02012B7C, (u32)dsiSaveOpen);
+		setBL(0x02012BFC, (u32)dsiSaveClose);
+		setBL(0x02012F9C, (u32)dsiSaveCreate);
+		setBL(0x02012FAC, (u32)dsiSaveOpen);
+		setBL(0x02012FC0, (u32)dsiSaveSetLength);
+		setBL(0x02013008, (u32)dsiSaveClose);
 		*(u32*)0x02013180 = 0xE1A00000; // nop
 		*(u32*)0x020148E8 = 0xE1A00000; // nop
 		*(u32*)0x020148F0 = 0xE1A00000; // nop
